🎨 extract mongo_curve_backend.py
